fix(Deadlock): Resolve a deadlock when updating state which is dispatched to Atomic Property Wrapper (#4492)
Thanks to @thisisalexmcgregor for finding this issue and debugging it with me.

After we update session state properties we [dispatch updates to all subscribers via their subjects](https://github.com/blockchain/wallet-ios-private/blob/dev/Modules/BlockchainNamespace/Sources/BlockchainNamespace/Session/Session%2BState.swift#L369-L376), this happens synchronously so that tracking the stack trace, testing and much more is easier to reason about. 

The bug, internally we use NSRecursiveLock to ensure safe access across threads - there is one common case where this can cause a deadlock. If you take the lock on a non-main queue, dispatch_sync to the main queue, and then try to take it again it will lock. [We have a wrapper called `Atomic` which calls `sync`](https://github.com/blockchain/wallet-ios-private/blob/dev/Modules/Tool/Sources/ToolKit/Utilities/Atomic.swift#L15) caused by one of our updates - this triggers the race condition.

This fix tightens where we actually synchronise access to the mutable properties and does not require the work from the subject dispatch to finish before we release the lock.

Attached is a snapshot of Alex's debugger at the time showing the two threads causing the deadlock 

![image](https://user-images.githubusercontent.com/88668746/186740323-db9626c6-00ea-4dd8-b815-10c60afd0a8a.png)

## How to Test

I was not able to reproduce this, however Alex was - to do this he was opening the app and triggering a pin login. This only started happening recently so is likely tied to changes made in either Native Wallet or Super App.