fix(Deadlock): Lock on Login (#4519)
## Description

When dispatching updates from a synchronised context we have to be extremely careful not to switch threads, this fix avoids calling update before releasing the lock. 

In the previous fix, I missed this case, this applies the same logic we did previously in another place. I vetted all cases inside session state and can confirm there should be no more cases of this bug present.

## Merge Checklist

- [x] The PR uses a title supported by [.changelogrc](https://github.com/blockchain/My-Wallet-V3-iOS/blob/master/.changelogrc).
- [x] Code is formatted using `sh scripts/format-code.sh`.