feat(CC Optimization): IOS-7025 - Adds feature flag support to CC opt… (#4506)
## Objective

When users go through the buy flow, there's an element of card optimization work there. It shows up in two spots:
1. Payment selection screen
2. The black pill button that stops the user from completing a buy order

This feature flag is the same flag that is used in the `Add a Card` flow. If that flag is disabled, warnings about cards that may not work or wont work at all will not show. 

## How to Test

Disable the feature flag. See that buy behaves the same way as before and you can buy with any card. 

## Merge Checklist

- [x] The PR uses a title supported by [.changelogrc](https://github.com/blockchain/My-Wallet-V3-iOS/blob/master/.changelogrc).
- [ ] Code is formatted using `sh scripts/format-code.sh`.
- [ ] UI components have accessibility identifiers and labels to support automated testing.
- [ ] You have added unit tests if applicable.
- [ ] You have added sufficient documentation (descriptive comments), if applicable.
- [ ] If a user facing feature, you have posted a Loom to the appropriate Slack channels.

Co-authored-by: Paulo <paulo@blockchain.com>