fix(CoinView): IOS-0000 Remove Swap Button for legacy app mode (#4554)
## Objective

The logic for the swap button applies only for the superapp, when
CoinView has only one type of account (Trading or PKW). When we have
both types of accounts in there, weird things can happen as reported by
@thisisalexmcgregor .
To simplify things rather than add more complex logic that will be
deleted later on, we've decided to hide the swap button, keeping the
same behaviour as the current production version of the app.