fix(Warnings): IOS-0000: 2.0 - Reduce warnings to near zero (#4566)
## Objective

The current level of warnings in the app makes real warnings from Xcode
updates or SDK's hidden. Whilst this has not lead to a bug yet, it's one
waiting to happen. We should reduce our build-time warnings as close to
zero as possible.

## Description

In this P/R there are several elements tackled:

1. Fix deprecations on usage of the iOS SDK
2. Remove superfluous disable commands
3. Fix all linting warnings
4. Disable some ObjectiveC warnings, which will naturally go when we
remove js/legacy
5. Remove some unnecessary rules from linter

There are some deprecations about non-native Bitcoin, this may go with
the work @paulo-bc is doing:

![image](https://user-images.githubusercontent.com/88668746/189346304-02e57686-61ed-4798-b0d1-17757acdfd9e.png)

@dchatzieleftheriou-bc @paulo-bc I would appreciate insight into the
following deprecation warnings, do we have any alternatives? Are the
warnings still valid?

![image](https://user-images.githubusercontent.com/88668746/189346531-10c2e598-09d3-4be6-aa26-84e66c564f0e.png)


![image](https://user-images.githubusercontent.com/88668746/189346578-79945ca4-d561-4874-8de3-1c30a0894773.png)

With the work done on this P/R we have significantly less warnings than
before, a month ago we had nearly 2000 warnings and today we will have
about 40 left which will likely go

## How to Test

Open and build the app

## Screenshot

*If the change is UI related, paste a screenshot here and include an
assessment of design accuracy if applicable.*


## Merge Checklist

- [x] The PR uses a title supported by
[.changelogrc](https://github.com/blockchain/My-Wallet-V3-iOS/blob/master/.changelogrc).
- [x] Code is formatted using `sh scripts/format-code.sh`.

Co-authored-by: Paulo <paulo@blockchain.com>