fix(Analytics): Update Tag+Reference.swift to fail when passed an invalid identifier (#4591)
## Description

Analytics values being reported by the data-driven firebase map would
work for tagged values but not for static constants. This is because the
initialiser for Tag.Reference would always pass, even when the string
was not a valid identifier. This makes sure the initialiser fails and
the `Either` implementation falls back to use the String

## Merge Checklist

- [x] The PR uses a title supported by
[.changelogrc](https://github.com/blockchain/My-Wallet-V3-iOS/blob/master/.changelogrc).
- [x] Code is formatted using `sh scripts/format-code.sh`.