fix(Coin): Remove receiveAddress from CoinView, improve the time-to-load (#4595)
## Description

Paulo noted that we were fetching receive address without actually using
it, and that would prevent displaying any account information until all
accounts had loaded. This changes this so that we display account
information as-and-when it's ready. We know the account names and types
ahead of time so the animation is smooth for loading in. See video.


https://user-images.githubusercontent.com/88668746/190421893-b8676a19-f812-4f45-adca-80638e7fdef2.mp4

Note, this also changes the account list from `zip` to `combineLatest`
so that each account is able to update independently of one another

## How to Test

- Open Coin View
  - on your home wifi
  - on your cell network
  - on a slow network (simulated bad network, GPRS?)

## Merge Checklist

- [x] The PR uses a title supported by
[.changelogrc](https://github.com/blockchain/My-Wallet-V3-iOS/blob/master/.changelogrc).
- [x] Code is formatted using `sh scripts/format-code.sh`.