feat(KYC): IOS-6295 Refactor KYC's Personal Info Screen (#3923)
## Objective

Start revamping the KYC Flow's UI with the Personal Info screen.
([Figma](https://www.figma.com/file/IXwonpgbc67ho6uvPX7O74/NUX---New-User-Experience?node-id=3822%3A59254))

## Description

I've used `FeatureForm` to implement this. That module is already used
for the account usage questions in the KYC flow, but the contents is
simpler and fetched from an API. For this screen, I've expanded the
capabilities to display date fields and allow each answer to be custom
validations. Eventually, we'll be able to use `FeatureForm` to power the
entire KYC flow with different questions set and order depending on the
user's country, state, etc.

## Key Classes

- `PersonalInfoView`
- `PersonalInfoReducer`
- `FormAnswerView`

## How to Test

*Describe how a reviewer can verify your changes. This helps with those
who are unfamiliar with the feature.*

## Screenshot

![Simulator Screen Shot - iPhone 8 - 2022-04-11 at 13 50
56](https://user-images.githubusercontent.com/82030034/162975164-8853ccca-7e9c-4257-aef6-7494e5f761b7.png)
![Simulator Screen Shot - iPhone 8 - 2022-04-11 at 13 50
58](https://user-images.githubusercontent.com/82030034/162975168-7e2be292-4c9e-423c-a130-e5638423d81d.png)

## Merge Checklist

- [x] The PR uses a title supported by
[.changelogrc](https://github.com/blockchain/My-Wallet-V3-iOS/blob/master/.changelogrc).
- [x] Code is formatted using `sh scripts/format-code.sh`.
- [x] UI components have accessibility identifiers and labels to support
automated testing.
- [ ] Areas of technical debt are marked with a `// TICKET:` comment
that includes a ticket number.
- [x] You have added unit tests.
- [ ] You have added sufficient documentation (descriptive comments).
- [x] New files added are within the correct directory. (e.g. if a file
is required for unit tests to compile, be sure it is added to the tests
target.)
- [x] ~If user facing feature: I have posted to #team-blockchain and
#mobile-product-updates, link: <LINK>~ Will do after the entire flow's
UI is revamped.

Co-authored-by: Oleh Kudinov <obrusentsev@blockchain.com>