fix(NativeWallet): IOS-5400 Wallet Payload Replenishments (#4753)
## Objective

[JIRA Ticket Link](https://blockchain.atlassian.net/browse/RI-866)

Some customers reported seeing decoding errors and can't login. 
This is due to variations of wallet payload and missing fields. 

In order to provide a better experience we need to replenish/recreate
these broken accounts.

## Description

This makes the decoding less strict but adds checks to see if the
`Account`, `Derivation`s need replenishment.
This also fixes an issue with syncPubKeys failing on wallet sync.

## How to Test

Given the nature of the issue we need to find broken accounts, which in
order to get such accounts we need to intentionally create broken ones.

I have tested this with manually removing (`xpriv`, `xpub`) properties
from my prod account and it fixed the missing fields.
Also performed some additional tests on a staging account with pwk
funds.

You can actually partially test this if you want
- Create an account from current dev
- Switch to this branch
- Build & Run
- Login, create or enter pin
- Verify Dashboard loads correctly

## Merge Checklist

- [ ] The PR uses a title supported by
[.changelogrc](https://github.com/blockchain/My-Wallet-V3-iOS/blob/master/.changelogrc).
- [ ] You have added sufficient documentation (descriptive comments), if
applicable.

Signed-off-by: dchatzieleftheriou-bc <dchatzieleftheriou@blockchain.com>